version 1.2
Numerous bug fixes: fixed passing in rows and columns as objects (previously resulted in
an empty dataset as the object would be pushed to an array), fixed rows and columns event
handlers not being called because of namespacing (namespacing of event handlers is now
user controlled), fixed column names if columns is passed as an object, fixed column
name if array is passed (now uses a "one" based index instead of "zero").

Added convenience methods for triggering event handlers and error handlers, useful
especially when called from outside the class.  Added a "filters" parameter to the
load function which allows you to apply filters to a single request only (does not
store these filters in class).